# docs/admin.openapi.yaml
...
openapi: 3.0.3
info:
  title: MetroMatrix admin console API
  version: 2.0.0
  description: >-
    Every admin endpoint (/api/admin, /api/v1/admin, /api/shopping/admin) answers in one envelope:


    - success: `{ success: true, data, meta? }` — `meta` pages lists (`page, limit, total, pages, nextCursor`); `limit`
    is clamped to 1–100.

    - failure: `{ success: false, error: { code, message, details? }, requestId }` — clients branch on `error.code` (see
    src/utils/errorCodes.js).


    Authentication: `Authorization: Bearer <accessToken>` from POST /api/admin/auth/login. Access tokens are
    short-lived; renew with POST /api/admin/auth/refresh-token (rotates the refresh token — replaying an old one revokes
    the session).

    A session may be restricted (`restrict: password_change | totp_enrol`) until the admin fixes their account.


    `x-guards` lists the permission guard of each operation (source: docs/ROUTES.json). Money, personal data and audit
    reads are guarded too.

    Status lists, labels, tones, limits and reference data come from GET /api/admin/meta — never hardcode them in a
    client.
servers:
  - url: /
    description: API host (Vercel)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationFailed:
      description: Validation failed (details.fields lists the problems)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Not signed in, token expired, or session revoked/idle
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Missing permission, super admin required, or session restricted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        requestId:
          type: string
          nullable: true
...
    SettingsResponse:
      type: object
      required:
        - values
        - spec
      properties:
        values:
          type: object
          additionalProperties: true
        spec:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        lastUpdatedBy:
          type: string
          nullable: true
...
paths:
...
  /api/admin/settings:
    get:
      tags:
        - Settings
      operationId: getAdminSettings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: '#/components/schemas/SettingsResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
...